remove reducers
