--- src/actions/advertisers/index.ts
@@ -1,2 +1,0 @@
-export * from "./advertisers.get";
-export * from "./advertisers.interface";
