--- src/actions/auth/index.ts
@@ -1,2 +1,0 @@
-export * from "./auth.interface";
-export * from "./auth.signin";
