--- src/actions/index.ts
@@ -1,3 +1,0 @@
-export * from "./auth";
-export * from "./advertisers";
-export * from "./snackbar";
