--- src/actions/snackbar/snackbar.ts
@@ -1,11 +1,0 @@
-export const OPEN_SNACKBAR = "OPENSNACKBAR";
-export const OpenSnackBar = (payload: any) => ({
-  payload,
-  type: OPEN_SNACKBAR,
-});
-
-export const CLOSE_SNACKBAR = "CLOSESNACKBAR";
-export const CloseSnackBar = (payload: any) => ({
-  payload,
-  type: CLOSE_SNACKBAR,
-});
