+++ src/components/EnhancedTable/renderers.tsx
@@ -18,8 +18,8 @@ import {
 } from "../../graphql/ad-set.generated";
 import { OnOff } from "../Switch/OnOff";
 import { CreativeFragment } from "../../graphql/creative.generated";
-import { IAdvertiser } from "../../actions";
 import { CampaignSource } from "../../graphql/types";
+import { IAdvertiser } from "../../auth/context/auth.interface";
 
 export type CellValueRenderer = (value: CellValue) => React.ReactNode;
 const ADS_DEFAULT_TIMEZONE = "America/New_York";
