--- src/reducers/advertiser/advertiser.interface.ts
@@ -1,5 +1,0 @@
-import { IAdvertiserPayload } from "../../actions";
-
-export interface IAdvertiserState {
-  advertisers: IAdvertiserPayload[];
-}
