--- src/reducers/auth/auth.interface.ts
@@ -1,8 +1,0 @@
-export interface IAuthState {
-  accessToken: string;
-  email: string;
-  id: string;
-  signedIn: boolean;
-  role: string;
-  emailVerified: boolean;
-}
